Move output format-deciding function to cli.go
